no size in bytes

+++ pkg/handlers/recursor/handler.go
@@ -22,7 +22,7 @@ type handler struct {
 	recursor *recursor.Recursor
 }
 
-func NewRecursorHandler(next pigdns.Handler, datadir string, cacheSize int64) *handler {
+func NewRecursorHandler(next pigdns.Handler, datadir string, cacheSize int) *handler {
 	h := &handler{
 		Next:     next,
 		recursor: recursor.New(datadir, cacheSize),
