+++ pkg/recursor/anscache.go
@@ -20,7 +20,7 @@ type ansCache struct {
 }
 
 // this is a cache for answers in the dns.Msg form
-func newAnsCache(datadir string, name string, size int64) *ansCache {
+func newAnsCache(datadir string, name string, size int) *ansCache {
 	rc := &ansCache{
 		cache: cache.NewFileCache(datadir, name, size),
 		name:  name,
